Extracted animation controller
